chore(website): fix metadata
